OSINT/requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
dnspython==2.3.0
nslookup==1.7.0
pygeoip==0.3.2
requests==2.28.2
soupsieve==2.4
whois==0.9.27
